+++ src/include/Camera.h
@@ -8,27 +8,27 @@
 #include "ZBuffer.h"
 
 
-using Vector3f = Eigen::Vector3f;
-using Vector4f = Eigen::Vector4f;
+using Vector3d = Eigen::Vector3d;
+using Vector4d = Eigen::Vector4d;
 using Vector2i = Eigen::Vector2i;
-using Point3f = Eigen::Vector3f;
+using Point3f = Eigen::Vector3d;
 using Json = nlohmann::json;
 
 class PinHoleCamera {
   public:
-   PinHoleCamera(Point3f lookFrom, Point3f lookAt, Vector3f up,
-                 float xFov = 45.f, Vector2i resolution = Vector2i(512, 512));
+   PinHoleCamera(Point3f lookFrom, Point3f lookAt, Vector3d up,
+                 double xFov = 45.f, Vector2i resolution = Vector2i(512, 512));
    PinHoleCamera(const Json& cameraJson);
-   Vector3f lookAt, up, right;
+   Vector3d lookAt, up, right;
 
   public:
-   Matrix4f view, projection, viewPort;
-   Matrix4f worldToScreen;
+   Matrix4d view, projection, viewPort;
+   Matrix4d worldToScreen;
    std::unique_ptr<RGBColorImage> film;
    std::unique_ptr<ZBuffer> zBuffer;
 
   private:
-   void setActualValue(Point3f lookFrom, Point3f lookAt, Vector3f up,
-                       float xFov, Vector2i resolution);
+   void setActualValue(Point3f lookFrom, Point3f lookAt, Vector3d up,
+                       double xFov, Vector2i resolution);
 };
 #endif
